requirements.txt
tensorflow
fastapi
uvicorn
numpy
pydantic
nltk
pymongo
python-dotenv
pymongo[srv]